Move character generation to it's own folder
