+++ toolkit/views/character_generator/character_generator.py
